Move database encoding helpers.
