+++ database/database.go
@@ -8,7 +8,6 @@ import (
 	"context"
 	"crypto/ed25519"
 	"crypto/rand"
-	"encoding/binary"
 	"fmt"
 	"io"
 	"net/http"
@@ -87,38 +86,6 @@ func writeHotBackupFile(db *bolt.DB) error {
 
 	log.Tracef("Database backup written to %s", backupPath)
 	return err
-}
-
-func int64ToBytes(i int64) []byte {
-	bytes := make([]byte, 8)
-	binary.LittleEndian.PutUint64(bytes, uint64(i))
-	return bytes
-}
-
-func bytesToInt64(bytes []byte) int64 {
-	return int64(binary.LittleEndian.Uint64(bytes))
-}
-
-func uint32ToBytes(i uint32) []byte {
-	bytes := make([]byte, 4)
-	binary.LittleEndian.PutUint32(bytes, i)
-	return bytes
-}
-
-func bytesToUint32(bytes []byte) uint32 {
-	return binary.LittleEndian.Uint32(bytes)
-}
-
-func bytesToBool(bytes []byte) bool {
-	return bytes[0] == 1
-}
-
-func boolToBytes(b bool) []byte {
-	if b {
-		return []byte{1}
-	}
-
-	return []byte{0}
 }
 
 // CreateNew intializes a new bbolt database with all of the necessary vspd
