+++ database/helpers.go
@@ -5,6 +5,7 @@
 package database
 
 import (
+	"encoding/binary"
 	"encoding/json"
 )
 
@@ -25,4 +26,36 @@ func bytesToStringMap(bytes []byte) (map[string]string, error) {
 	}
 
 	return stringMap, nil
+}
+
+func int64ToBytes(i int64) []byte {
+	bytes := make([]byte, 8)
+	binary.LittleEndian.PutUint64(bytes, uint64(i))
+	return bytes
+}
+
+func bytesToInt64(bytes []byte) int64 {
+	return int64(binary.LittleEndian.Uint64(bytes))
+}
+
+func uint32ToBytes(i uint32) []byte {
+	bytes := make([]byte, 4)
+	binary.LittleEndian.PutUint32(bytes, i)
+	return bytes
+}
+
+func bytesToUint32(bytes []byte) uint32 {
+	return binary.LittleEndian.Uint32(bytes)
+}
+
+func bytesToBool(bytes []byte) bool {
+	return bytes[0] == 1
+}
+
+func boolToBytes(b bool) []byte {
+	if b {
+		return []byte{1}
+	}
+
+	return []byte{0}
 }
